move common container check to utils.go
